+++ auth-db/queries/users/index.ts
@@ -1,2 +1,3 @@
 export * from "./create.ts";
 export * from "./get-all.ts";
+export * from "./get-by-email.ts";
